ミス修正とリファクタちょっと
